+++ service-tree-navigator-example/src/main/java/com/zhuinden/servicetreenavigatorexample/FirstView.java
@@ -1,0 +1,60 @@
+package com.zhuinden.servicetreenavigatorexample;
+
+import android.annotation.TargetApi;
+import android.content.Context;
+import android.support.annotation.Nullable;
+import android.util.AttributeSet;
+import android.view.View;
+import android.widget.LinearLayout;
+
+import com.zhuinden.navigator.Navigator;
+import com.zhuinden.servicetreenavigatorexample.injection.FirstComponent;
+import com.zhuinden.simplestack.Backstack;
+
+import butterknife.ButterKnife;
+import butterknife.OnClick;
+
+/**
+ * Created by Owner on 2017. 03. 17..
+ */
+public class FirstView
+        extends LinearLayout {
+    public FirstView(Context context) {
+        super(context);
+        init(context);
+    }
+
+    public FirstView(Context context, @Nullable AttributeSet attrs) {
+        super(context, attrs);
+        init(context);
+    }
+
+    public FirstView(Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
+        init(context);
+    }
+
+    @TargetApi(21)
+    public FirstView(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
+        init(context);
+    }
+
+    private void init(Context context) {
+        if(!isInEditMode()) {
+            FirstComponent firstComponent = Services.getNode(Backstack.getKey(context)).getService(Services.DAGGER_COMPONENT);
+            firstComponent.inject(this);
+        }
+    }
+
+    @OnClick(R.id.first_button)
+    public void clickFirst(View view) {
+        Navigator.getBackstack(view.getContext()).goTo(SecondKey.create());
+    }
+
+    @Override
+    protected void onFinishInflate() {
+        super.onFinishInflate();
+        ButterKnife.bind(this);
+    }
+}