+++ service-tree-navigator-example/src/main/java/com/zhuinden/servicetreenavigatorexample/Key.java
@@ -1,15 +1,13 @@
 package com.zhuinden.servicetreenavigatorexample;
 
-import android.support.annotation.LayoutRes;
-
+import com.zhuinden.navigator.StateKey;
 import com.zhuinden.servicetree.ServiceTree;
 
 /**
  * Created by Owner on 2017. 03. 17..
  */
 
-public interface Key {
-    @LayoutRes int layout();
-
+public interface Key
+        extends StateKey {
     void bindServices(ServiceTree.Node node);
 }